Accounts/pyproject.toml
[tool.poetry]
name = "accounts"
version = "0.1.0"
description = ""
authors = ["Harmit <harmit.singh@live.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.2.2"
flask-httpauth = "^4.7.0"
pyjwt = "^2.6.0"
flask-sqlalchemy = "^3.0.3"
flask-restful = "^0.3.9"
pymysql = "^1.0.2"
python-dotenv = "^0.21.1"
flask-restplus = "^0.13.0"
marshmallow-sqlalchemy = "^0.29.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
